requirements.txt
langchain>=0.1.0
langgraph>=0.0.27
langchain-google-genai>=0.0.5
pydantic>=2.5.2
python-dotenv>=1.0.0
typing-extensions>=4.8.0
pytest>=7.4.3
